Add user description controller actions and view
This adds two controller actions

- GET :user_description
- POST :confirm_user_description

to the early access entry controller. It adds a view for the user description
endpoint.

The POST endpoint adds the users choice of description to the
session["sign_up"] hash.

I've had to add some margin to the bottom of the backlink to stop the error
summary looking funky.