Add reason for visit form
This handles taking the eligibility question choice and validating it's
presence. If valid then it creates the new EarlyAccessUser, using the
choice selected as the reason for visit.

It takes two additional attributes, email and user description. I did
think about adding validations for these but decided against it as they
will always be passed into the form in the controller.

If the values not present in the session then the user will be redirected
to the correct endpoint so they won't ever not be present in the form.

Once the user is created it creates a passswordless session and emails
the user the magic link.