Ensure that reason for visit endpoint cannot be reached in invalid state
At this point the sessions sign_up should containt he users email address
and reason for visit.

If that isn't the case then they need to be redirected to the initial
page of the sign up flow if email is missing and the user description
page if email is present but description is missing.