Add shared examples and contexts for early access request specs
There's a tonne of duplication going on in these tests now some additional
tests have been added for the reason for visit endpoints.

This adds a couple of shared_contexts for when the user has provided
their email address and when they've provided their email and description.

It also pulls out the tests that check the correct redirects occur
when the session[:sign_up] doesn't contain the correct information
into shared_examples.