Add reason for visit controller actions and view
This adds the controller actions and view for the reason for visit page.
It's view is very similar to the user description page.

The confirm_reason_for_visit action either:

- re-renders the reason_for_visit page if they've not made a choice
- submits the form, clears the sign up session and renders the email
sent page if they have made a choice

I've also added some system specs since the flow is now complete.