Clean up of OpenAI chat stubs
This starts the process of having more consistency between the OpenAI
chat completion stubs and simplifying the amount that needs to be
stubbed.

The motivation for this was changing models earlier today and having to
do shotgun surgery across the codebase to fix tests broken by it.

The main changes here are:

- moving the question_or_history from the structured answer stub to be
  available across all the stubs that accept questions
- gone consistently for a_string_including for matching string
  intersections
- asserting model by a regex and using a hash_including so we don't need
  to pass everything into the OpenAI parameters for every assertion
- replacing the static methods in the stub module with verbosely named
  instance methods, this allowed use of RSpec matchers and was more
  consistent with other helper functions.

The only negative consequence I've found from this change is that for
doing partial matches of answers (e.g regex or a_string_including) we
need to be using an RSpec matcher, like array_including, otherwise they
fail to match ☹️.