Make answer a kwarg on OpenAI chat stubs
We set this as a required argument when we expected all our OpenAI stubs
to involve a message as part of an answer. Now that we're using tool
calls we don't necessarily have a message with content and we have a few
confusing method calls where this is set to nil.