Switch sidekiq concurrecy to 25
This sets the concurrency value to be the default Sidekiq one of 25.
The motivation for this is that I wanted to try get the highest level of
concurrency for the red teaming exercise next week so we don't delay
answers to users.

However, assuming our pods have sufficient resources for this which I
believe they do, I can't see any harm of this value.