Control sidekiq concurrency with an env var
It turned out that 25 exceeded the Redis connection limit on our Heroku
mini redis instances, so this allows Heroku to be configured differently
to other environments.

If it turns out that 25 is too much for local machines we may also need
to break this file into a development and production section - or just
tweak the env var in GOV.UK Helm charts.