Handle users who try to sign up and already have an account
It's pretty unlikely that a user will reach the reason for visit page
when they already have an account, but it is possible.

For example, two people who share an email address start signing up at the
same time. Or one person could start signing up one device, finish signing up
on another device, then return to the first device.

This would cause an unhandled exception. This updates the endpoint to
handle a conflict and render some plain text telling the user they've
already got an account.

It might be that we want to add a page in GOV.UK styling for this down
the line.