Have two lists for messages
This re-implements the spike work from
https://github.com/alphagov/govuk-chat/pull/441 and https://github.com/alphagov/govuk-chat/pull/416

This work set-up two lists of messages for helping users that use a
screenreader.

The need for two lists is explained as:

"""
I've seen an issue where VoiceOver stops announcing new messages or cuts off announcements after a certain point. I believe this may be due to internal thresholds VoiceOver has for the amount of content it reads from a live region, to maintain responsiveness and usability.

To address this, I've created a separate, non-live container where messages are moved before new ones are added to the conversation. This keeps the amount of content in the live container minimal.
"""

In implementing this it wasn't 100% clear from the spikes what was
deliberate element choices and what was artefacts from spiking, so I've
tried to organise the same but with fixes for things that seemed
illogical (like two width containers) - hopefully is this wrong it'll
get caught in a review.

I've stripped the JS classes so that the JS side can be implemented
in later commits.

I've refactored the views since they have virtually no HTML in them any
more and can be just ERB.

Co-authored-by: Daniel Karaj <daniel.karaj@digital.cabinet-office.gov.uk>
Co-authored-by: Jon Kirwan <87758239+jon-kirwan@users.noreply.github.com>