EarlyAccessUser: validate unique email address
We should handle the scenario where an admin adds an EarlyAccessUser
with an email address that already exists.

This is handled at the database level with a unique index, but adding
this validation to the form handles it gracefully, rather than raising
an exception.

Initially I assumed it'd be a matter of extending the original
`validates` directive, like so:

```
validates :email, format: { with: URI::MailTo::EMAIL_REGEXP, message: "Enter a valid email address" },
                  uniqueness: { message: "Email address already exists" },
                  if: -> { email.present? }
```

but that resulted in this exception being thrown:

```
Unknown validator: 'UniquenessValidator'
```

I think it's to do with the uniqueness validator relying on the database
to enforce uniqueness and doesn't play nicely with non ActiveRecord
classes. So I did the validation manually instead.

This commit also changes the `email` ivar to be an attribute, to be
consistent with all the rest of the form models we have.