Allow editing of waiting list users
This updates the existing `CreateWaitingListUser` form to also handle
updating an existing user.

The form works on the basis that if the `user` attribute is present,
we're updating the user, otherwise we're creating a new user.

It also sets up the `edit` and `update` actions for the
`WaitingListUsersController` and wires up the form to allow editing of
the data.

It refactors the form into its own partial so it can be shared between
the `new` and `edit` views.

Form