Update OnboardingController#privacy_confirm to set onboarding_completed
When the user has completed both onboarding pages/steps we want to set the
onboarding_completed boolean to true.

I've gone with a more straightforward approach of using a boolean which rather
than an enum or something more complex as in reality almost noone will
complete the first onboaring page and not return and if they do it's
really not a hassle form their to click the button again.

Also they probably should re-read the page anyway.