Update sessions_controller to set cookies and session variables
It feels like the easiest way to ensure that the application works in
both environments it to retain the current behaviour for Heroku and to
handle setting the required session variables and cookies at the point
of sign in for production.

This updates the SessionsController#confirm endpoint to set the
session[:onboarding] and cookies[:conversation_id] when appropriate.

It does the following:

1. deletes the conversation_id cookie if present

We need to do this to protect against users seeing each others conversations
if they're using the same browser.

2. if cookies[:conversation_id] isn't found but the user has an active
conversation then it sets the cookies[:conversation_id] to the users latest
active conversations id

3. If the user has completed onboarding then it sets the session[:onboarding]
to 'conversation'

In each of these situations the user will be redirected to the onboarding
limitations page which will then redirect them to the conversation show
page.

If none of the above is true (normally the first sign in) then there will
be no behaviour change. They'll be redirected to the onboarding page
and complete the onboarding process as normal.