Add manual calls to focus message region
This replicates the behaviour from [1] and [2] where we manually call
focus on the message region to help screenreader users.

I understand these focuses are intended for just after we insert dom
elements into HTML so the user can be alerted of them.

There's a few inconsistencies around this that I didn't address and may
be caught in review: we don't set focus when we're progressively
disclosing existing messages, only when appending new ones - should we?
Similarly I understand we're not doing any focus change at question
submission just when answer is available, which seems a bit inconsistent
and may prove surprising.

[1]: https://github.com/alphagov/govuk-chat/pull/441/files#diff-731fb64fbe90b7dd33be95f9b7027271c58bf28ff965055561e79c08d8d8eb2bR123
[2]: https://github.com/alphagov/govuk-chat/pull/441/files#diff-347bc348b11debdd593b19cdd9e31415b3c2bd37c72a0baa3ed586f4b0c848b1R53