Re-introduce fade behaviour for new questions and answers
This sets up questions to either fade the loading message in or the
question depending on whether the loading message loads first - this is
done so that there is only one fade and it can look like the loaded
question replaces the loading message.

For an answer, we always expect answer generation to take longer than
the time for the fade so we just fade the answer loading message in and
never attempt the answer.

I believe this is roughly consistent with [1].

[1]: https://github.com/alphagov/govuk-chat/pull/441/files#diff-347bc348b11debdd593b19cdd9e31415b3c2bd37c72a0baa3ed586f4b0c848b1R53