Handle OpenAI returning a string in the response body for errors
We made the incorrect assumption that OpenAI always responds in the same
format when an OpenAIClient::RequestError is raised. It seemingly can
respond in 1 of 2 ways:

1. as a hash with the following format

```
response

{ body: { "error" => { "message" => "nested error message" } }}
```

2. as a string

```
respponse
{ body: "error message as a string" }
```

This updates the OpenAIAnswer class to handle the second case. It also adds
a test to ensure that if OpenAI responds with a hash in an unexpected format
we use the error message on the error object as default.