Extend chat_introduction component to cover early access landing page
We're goin to have 2 landing pages, 1 for heroku where the user doesn't
need an account and 1 on production where a user is asked to input their
email and go through a sign up flow.

This extends the introduction component to work in both cases. It's
a bit clunky as there's a few extra fields that need to be passed into
the input (error items and value), but there's so much shared styling
this felt like the lesser evil than a brand new component.

I ran into an issue with the current implementation when errors were
triggered. There was 100px applied to the wrapper which shoved it miles
down the page.

I've added an override class when errors are present which strips the
top padding. This leaves 40px of margin on the bottom of the error summary
component and looks decent to me. But it'll need a some FE eyes.

Also the reason i've used
'Rails.application.routes.url_helpers.early_access_entry_sign_in_or_up_path'
for the form url is because the component guide blows up without it.