Ensure early access radios link to radio fieldset
hidden_field_tag constucts a default id for the hidden field based on the
name of the field.

form[choice] would become form_choice. This meant that the error summary
component was linking to the hidden field instead of the radio fieldset
since both had the same id and the hidden field was rendered first.

This passes an explicit id to the hidden field to ensure that the errors
link correctly.