Refactor QuestionFilter to use BaseFilter
The `BaseFilter` class was added to encapsulate shared logic between the
`EarlyAccessUsersFilter` and `WaitingListUsersFilter`.

Quite a bit of the logic is also present in the `QuestionsFilter` class
too, so this commit makes that class inherit from the `BaseFilter`.

To make the filters more generic, the `users` and `questions` methods
have been renamed to `results`. This allows all filters to use the
pagination logic from the base filter, which previously required the
inherited filtes to implement a `users` method which obviously doesn't
make sense for a question filtering class.

The `QuestionFilter` has been moved into the `admin/filters` directory
too, so that all the filters are grouped in the same place.