Redirect early access users who are onboarded to the conversation show page
We have some logic in place to ensure that users who haven't been onboarded
are redirected to the onboarding page. Currently, the logic checks for
a conversation cookie or that session[:onboarding] is set to "conversation".

This extends these before actions to include an early access users having
onboarding completed so to true.

It also extends/adds shared examples to test this behaviour.