Ensure we don't create duplicate users in the reason for visit form
If a user makes it to the reason to visit page and for whatever reason
already has an account or is on the waiting list we should not create a
duplicate user.

This updates the form to raise an error which is caught and handled in the
controller.

If the users on the waiting list we tell them and if they've got an account
we tell them and link them to the sign in page.