Update sign in or up endpoint to handle user being on waitlist
When the user is on the waitlist and inputs the associated email address,
we should render a page which informs them they're already on the waitlist.

This updates the sign in or up form to pass back a result object with
an outcome of :existing_waiting_list_user when the user is on the waitlist.

When this occurs we render a view which informs the user they're already
on the waitlist.