Defeat the progressive disclosure delay for working tests
As the progressive disclosure takes 5-10 seconds to reveal a form this
currently makes tests fail as they timeout waiting for the button. If we
were to wait for the button the test suite would be very slow. Therefore
I've set-up a mechanism that we can configure the delay to 0 for the
test environment.