Prevent bug causing dual progressive disclosure
This change resolves a bug that was introduced by the
progressiveDisclosure code, to always have a delay. This bug meant that
when we were initialising an Onboarding module without new messages and
then injecting new messages in, we had both of these calling intervals
on the same elements (which created the effect of pairs of messages
being shown).

This is resolved by checking for new messages on init before starting
the progressive disclosure delay.