Move conversation and onboarding controller behind auth
This commit adds the require_early_access_user! as a before action to the
base controller then skips it for all controllers other than the conversation
and onboarding controller.

It removes the ProtectedController which no longer has a role now that a
user is required for most pages.

Co-authored by: kevindew <kevindew@me.com>