Add static pages
This adds in pages for: about, support and accessibility. The first two
of these have content from the Content Designers, whereas there is just
dummy content for the accessibility page.

I've set these pages as cached for 5 minutes and skipping session so
they can be cached. I'm going to expect these are available even when a
user is not logged in.

The designs for these did originally include a "Back" link, however
following a conversation with design about where this should go - since
there is no clear proceeding pages - we've agreed to drop these.

For the privacy page it looks like this may remain on GOV.UK separately -
I've set up a redirect for this as it seemed good for us to still have a
/chat/privacy URL.

For links with target=_blank, I have set noreferrer for any links that
going outside GOV.UK services and noopener for those that are. The use
of the use of noopener is likely superfluous, but it is required by the
linter [1] despite being debatably outdated [2].

Edit: Following review I've added noopener as well to noreferrer links
to meet Design System guidance [3]. I've not been able to find an
example of a browser that needs this, let alone one in use - so I expect
this is unnecessary.

Co-authored by: Pete Goddard <peter.goddard@digital.cabinet-office.gov.uk>

[1]: https://github.com/rubocop/rubocop-rails/blob/master/lib/rubocop/cop/rails/link_to_blank.rb
[2]: https://github.com/rubocop/rubocop-rails/issues/1294
[3]: https://design-system.service.gov.uk/styles/links/#opening-links-in-a-new-tab