Move sign_in_or_up endpoints into landing page
This moves the landing pages for both flows into the same controller to
reflect them becoming the homepage. It required quite a lot of shotgun
surgery and did reveal a few bits that were a little broken.

I'm going to rename the chat controller in the follow up commmit to
HomepageController to reflect it's purpose.

Co-authored by: kevindew <kevindew@me.com>