Waiting list users: validate user research attributes
This commit validates the form when creating a waiting list user so that
the user reasearch fields (`reason_for_visit` and `user_description`)
are valid according to the enum we specify.

It also tidies up the test a bit because it was becoming very repetitive
to add these two attributes in on every spec.