Waiting list user form UX improvements
This sets the `error_id` on the two select fields for the user research
fields. This highlights them in red if the form contains an error
relating to them.

It also sets the `selected` option on each `option` so that if the user
selects an item and submits the form but the form is invalid, then we
set the option they originally chose.