Test MailRecipientInterceptor and fix issues
I found when testing out our mail interceptor that it was not actually
producing the prefix result that was expected. I imagine this is because
the code we copied for this [1] predates the use of the mail-notify gem
[2].

I've wrote a couple of tests for this now. Given the rather integrated
nature of this class I felt it was best to write the test as an
integration test rather than a unit test, as I didn't think we'd have
much value from a unit test on it's own and that an integration test
would cover all the cases of a unit test.

I still need to confirm this behaves by doing an integration deploy.

[1]: https://github.com/alphagov/content-publisher/blob/ec2b18596975ab6782c9614c7ac08216f1ef585f/lib/mail_recipient_interceptor.rb
[2]: https://github.com/dxw/mail-notify