Handle JSON responses to conversation missing
We shouldn't respond with a redirect to JSON responses unless that is a
valid response to the user request - as our JS code will parse follow
the redirect and try use the resulting JSON. This can produce niche bugs
in the system (I find myself fighting a system test that was setting an
answer as "undefined").

This applies a JSON response to the situations where a conversation
isn't found.

I also noticed that we didn't really test many of the conditions of this
in all the scenarios so I set up a shared example to be exhaustive. It
is all a bit ugly and does feel a bit over-the-top but does seem better
than having some scenarios just not tested.