Allow a waiting list user to unsubscribe
This adds two new routes to allow a waiting list user to unsubscribe.

There's a GET route which will render a simple confirmation page. If the
token in the URL params doesn't match any user record's
`unsubscribe_token` then it just redirects to the `chat_path`.

If the token matches, and the user submits the confirmation form, we
make a POST request to the confirm action which destroys the user record
and again redirects to `chat_path`.