Return JSON responses for onboarding flow issues
This change has been applied to resolve a potential bug scenario where
using redirects to correct an onboarding flow issue would cause problems
if done by JS. This is because the JS will follow the redirect and try
parse the result.

Returning bad request allows the JS to fall back to it's error handler.