JSON responses to user not signed in
Following the same precedent as the previous commits this adds in JSON
handling for when a user is not signed in. This is to ensure we get an
error response back when making JSON requests so that the JS experience
of the application does not break by following redirects.

I removed the inverse option from the shared example as I felt this
would have been fully covered by other tests for the endpoint.