Add method to determine whether question limit is reached
If the `question_limit` attribute is `0`, it means a user can ask as
many questions as they want (useful for internal testing).

If the value is `null`, then we apply the application default. Otherwise
we use the value from the field.

We use this value to determine whether the user has reached their limit
of questions.