Remove misleading back link
When a user reaches this page they have been through a number of forms,
thus you'd expect the back link to take you back to the most recent form
and not the homepage.

Going back to the homepage from here is also not ideal as this would
prompt a user to start generating a new magic link when they should have
just been sent one.

I need to confirm this change with design.