Fix footer on static pages
The footer was overflowing beyond its container due to the restrictions on the height property on the body element - setting the height to auto seems to resolve the issue. The same issue existed on the landing page so I initially added the .app-landing-page-layout class to the body element on static pages, which seemed to work well, but it didn't feel quite right to add the landing page layout class to pages that weren't the landing page, so I added an .app-static-page-layout class instead. Happy to discuss.