Mask whitespace beneath the footer
Whitespace beneath the footer was appearing on some screen resolutions because the animation boxes were extending beyond the footer, causing additional whitespace. I explored numerous options to try and address this, such as using clip-path, experimenting with the placement of the overflow rule, and setting a min-height on the main element to push the footer down to occupy whitespace when it appeared. These options either didn't work fully or didn't give me complete confidence that an edge case wouldn't crop up. Therefore rather than trying to remove the whitespace, I opted for masking it by setting the landing page's body element background to light grey. Short of rebuilding the animation so it works as required, I'm running out of ideas - happy to discuss further if an alternative would be more suitable.