Remove h2 in footer
As per the DAC audit, the h2 in the footer has been removed since it was not sufficiently descriptive. DAC suggested including a visually hidden h2 with the content 'Support links', which aligns with the design system. In order to get the visually hidden h2, you have to pass the links into the footer component via the meta option - not via the navigation option as had been used here. Therefore I've used the meta option to pass links to the footer, which has had the added benefit of allowing a chunk of override styles to be removed.