Remove redundant authenticity token
As this component is loaded on a cached resource we can't generate a new
token for each request. We do already ignore the value for this by
skipping the validation of a token, however I chose to remove this as
I'm experimenting with sessions and this was triggering a session
renewal.