Alter the header link based on auth status
We don't want users to be going back to the chat homepage when they are
logged in as that is a cached page that invites the user to log back in.

So instead we use their login status to decide whether the logo takes
the user back to their conversation or homepage based on authentication
status. For a user that hasn't yet got a conversation this link would
automatically redirect them to onboarding.

A consequence of this change is that I have removed the skipping of
session for static pages as we now vary the contents of the page based
on if there is a session or not.

I'm not sure if we'll have issues setting cache headers on the static
page with the CDN caching ones based on a user session. My assumption is
that the CDN won't cache anything that returns a cookie but I'll need to
confirm that once this app is behind the CDN, I think we'll likely need
to add extra CDN config to set this up.