Allow admin filtering of questions when scoped to a user
An admin can click on the question count for a particular user to be
shown just the questions that user has asked.

That filtering uses the same controller/filter as the normal question
filtering, except that we've added the `user_id` parameter to the URL so
the questions are filtered by that ID.

It's useful for an admin to then filter that user's questions using the
filters on the left hand side of the page (e.g. message, status, created
date and so on).

This commit extends the functionality to allow for that. It's a bit
different to the regular filters because we don't want an input field
for filtering users, i.e. a massive select field with every user we have
on the system.

The only way to get to a given user's questions is to go via their show
page and click on the question count.

So to allow additional filtering, we just render a hidden input field
with a name of `user_id` if the user ID is present in the URL params.
That way we're always scoping the other filters by that user's
questions too.

To make it obvious that the questions are being filtered by the user,
there's a bit of text at the top of the left hand pane saying "Filtering
by user: {email}", because otherwise you wouldn't know this filter was
in place. Clicking the "Reset all filters" will remove this filter along
with any other ones.