Remove conversation early_access_users foreign key
We want to be able to delete an early access user record but keep the
`early_access_user_id` field stored on the `conversations` table in
place.

This will allow us to delete a user but still be able to group questions
together for a particular user.

Removing the foreign key constrains allows us to keep the ID in place
even if the associated record isn't present.