Use absolute URLs in footer
There is a surprising feature of the layout_footer component that
converts any relative links used in the component into absolute ones
[1].

This feature breaks the footer links for GOV.UK Chat in any places where
chat is not run on www GOV.UK hostnames.

To counter it I've set absolute URLs as the data we input into the
component.

As we don't have precedent to test this view I've not added any tests,
as they'd also seem unprecedented. However if we find this issue
re-occurring by someone not realising why we have these we may have to
think about regression tests.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/3699