Waiting list: one-click unsubscribe
It's been decided that "unsubscribing" from the waiting list should be a
one-click thing, rather than first taking the user to a confirmation
page where they click a button to actually confirm.

This removes the `POST` endpoint and essentially moves that logic into
the `GET` endpoint instead.

We need to return early for `HEAD` requests as typically email clients
will make this request and we don't want to unsubscribe people when all
they've done is open the email.