Show deleted user's details on question show page
We use this helper method to generate the data we display on the question
show page.

Currently it just shows the user's details if the user is still present
in the database.

Now that we're allowing users to delete their accounts, we want to make
this more clear on that page. So instead of showing the user's email and
a link to their show page, we just show the text "Deleted user" instead.

We additionally show a link to view all of that user's questions,
regardless of whether they're deleted or not.