Show deleted user's details when filtering questions
Currently when filtering questions in the admin panel by a particular
user, we show a hint in the sidebar because there's no actual filter
input field for users.

We only do this if the user hasn't been deleted though. This commit
changes it so that we show the user's ID and "(Deleted user)" if the
account has been deleted.