Rename conversation view components for clarity
The conversation view is used for both messaging and onboarding. The current class names are confusing.

This commit splits the view into a `message-region` and a `form-region`. It uses `container` for `div` below that that wrap content and renames the `conversation-form` to `question-form` to more clearly differentiate it from the `onboarding-form`.