Remove nulls from JSON before exporting to BigQuery
We had an issue where BigQuery was rejecting our daily import because a
value had changed datatype and so wasn't compatible with the existing
schema.

We purposefully allow BQ to automatically detect our schema rather than
explicitly defining it so it's easier for us to add fields to the
export. We can just add a field to the JSON and it'll be pulled in
automatically.

However we started seeing this error:

```
Provided Schema does not match Table chat_export.questions. Field answer.question_routing_confidence_score has changed type from FLOAT to STRING
```

After a bit of digging we found that BQ treats `null` values in JSON a bit
strangely[1]. Essentially if it encouters a `null` value, it'll convert
that to the string `"null"`, hence why it was complaining that this
field went from a float to a string. This only happens on non-string fields.

The consensus is that to fix this, you simply omit the field from the
JSON object, rather than setting its value to `null`.

So this commit recursively traverses the JSON object and removes any
`null` values before writing it to the temp file.

[1] https://cloud.google.com/bigquery/docs/json-data#json_nulls