Register prometheus metrics on initialisation
We need to register the prometheus metrics on the default client on
initialisation, so that we can increment the sign up metrics when a user
signs up.

This commit adds a new lib/metrics.rb file with a ".register" method
that registers the counter metrics for instant and waiting list sign ups.

It then updates the config/initializers/prometheus.rb file to call the
".register" method after initialisation.