Add Metrics.increment_counter method
This extends the Metrics module to include a method for incrementing a counter.
This works by checking that a metric with the given name exists.

If it does then it increments the and provides the source as a label.
If not then it notifies sentry and returns. It won't raise the error so
the user will not be disrupted.