Don't assert the internals of the component we're using
This test was failing because the underlying component code changed in a
minor way which is an indication that this was over testing, by actually
doing unit testing on code outside of our unit.

I have therefore simplified the code to only have a single assertion for
the nested component and asserted that the navigation just contains the
links we're injecting.

I also removed the print utility assertion as this wasn't relevant to
the options under test.