Fix SCSS mixed declaration deprecations
This fixes the warning below (and the multiple repetitions of it).

```
Deprecation Warning: Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an upcoming
version. To keep the existing behavior, move the declaration above the nested
rule. To opt into the new behavior, wrap the declaration in `& {}`.

More info: https://sass-lang.com/d/mixed-decls

    ┌──> app/assets/stylesheets/components/_chat-header.scss
176 │     font-size: 16px; // directly set the font size to prevent govuk-font mixins from reducing it on smaller screens
    │     ^^^^^^^^^^^^^^^ declaration
    ╵
    ┌──> ../../.rbenv/versions/3.3.4/lib/ruby/gems/3.3.0/gems/govuk_publishing_components-43.0.1/node_modules/govuk-frontend/dist/govuk/vendor/_sass-mq.scss
229 │ ┌         @media #{$media-type + $media-query} {
230 │ │             @content;
231 │ │         }
    │ └─── nested rule
    ╵
    app/assets/stylesheets/components/_chat-header.scss 176:3  @import
    app/assets/stylesheets/component-guide.scss 11:9           root stylesheet

```