Make admin and frontend signon methods distinct
This moves the GDS SSO code for frontend area to be only configured in the
BaseController for that area and the same for the admin area.

The motivation for this was to change the way the frontend before action
is defined so that it has a name. This is so that this action can be
skipped by the ErrorController which is currently erroring in some
scenarios.