Don't let an error on loading current user kill the layout
As we use this layout for rendering errors we should be cautious that
errors in rendering will kill the ability to have a nice error page.

The area that struck me as being at risk of this was looking up the
EarlyAccessUser since that involves db calls. Therefore I've put this in
a begin rescue so we can still render it.

I've taken the liberty to risk calling GovukError if there is a fail.
There is, of course, a chance that if this raises an error too we get an
ugly error page.

In doing this I decided to move the tests for the navbar link to an
application layout spec, as they felt a bit more specific for that than
for the downtime spec.