Don't automatically promote users when public access is disabled
This prevents the rake task from performing automatic promotions while
public access is disabled. This has been done to avoid a situation where
the app is taken down and then new early access users are still created
and they receive emails with links that fail.

We do still have the functionality to promote or create waiting lists
users via the admin area if the app is not publicly available. This
seems a more manageable risk as that is an explicit action, whereas this
task is implicit.