Show answer metrics in admin panel
Renders a summary list of all the metrics for a question's answer if
it's present.

The metrics are sorted alphabetically so they're always in the same
order. It'd be nice to sort by pipeline step but that would require
keeping a separate list of ordered pipeline steps so it seems like more
work than is necessary.