Remove dynamic label name
Originally, whenever a user submitted an erroneous input and received an error message, it was announced to screen readers via aria-live (when JS was available/enabled) or aria-describedby (when JS when unavailable/disabled). Due to issues with aria-live, however, I implemented a change when JS was available/enabled whereby JS would modify the <label> element to reference any error messages when they occurred to get them consistently announced (PR #385). At the time, I did also explore using aria-describedby to announce the errors however I was concerned that users may not wait to hear the entire announcement since the hint text is announced before any errors. This was identified as a WCAG A failure during the DAC audit, however, and DAC recommended following the design system and to reference error messages via aria-describedby. So in a nutshell, this commit reverts most of PR #385 and instead aligns the JS version with the non-JS version re announcing error messages via aria-describedby.