Move current time method to AnswerComposition module
Previously this method was stored on the
`AnswerComposition::Pipeline::Context` class where it sort of made sense
in that it was easier to call in all the pipline classes, and easier to
test, but it didn't _really_ belong on the context. Also it meant that
it couldn't be used in the `AnswerComposition::Composer` class.

This moves it to a static method on the `AnswerComposition` module where
it can be used in all classes.