Remove message returning from QuestionRephraser
We don't actually need to return a value from here, all that matters is
that we assign the value to `context.question_message`.

This removes the return value and updates a couple of tests that broke.
One asserts the return value which, as discussed, we don't need.

The other test is asserting that the correct data is sent in the OpenAI
prompt request. The test will fail in the `stub_openai_chat_completion`
call if something is wrong, so we don't need to actually assert anything
in the test itself. Rubocop doesn't like that though so the rule needs
to be disabled for that test.