Ensure feedback is pre-hydrated to nil for abort_timeout answers
When a question has timed out, we create an answer with a status of
abort_timeout. When the requests response type is JSON, we attempt to
pass the answer html back.

When we attempt to render the answer partial to a string, it blows up
since the feedback has not been loaded. So we see a strict load error.

This updates the Question#check_or_create_timeout_answer to return an
answer with feedback pre-hydrated to nil when an abort_timeout answer is
created.