Remove redundant argument from question summary list
The `summary_list` component takes this `edit` argument to show a link
at the top of the component. But it only shows that link if you also
pass a `title` attribute. We're not doing that so nothing is showing.

We already have a link to the conversation in the actual list items, so
this is serving no purpose.